Move SettingsRadioItemSmall to a separate file.
